SOF-1550 Refactor AssetFolderHelper to follow naming conventions
