+++ src/blueprints/tv2/action-factories/tv2-dve-action-factory.ts
@@ -36,7 +36,7 @@ import { Tv2UnavailableOperationException } from '../exceptions/tv2-unavailable-
 import { Tv2CasparCgTimelineObjectFactory } from '../timeline-object-factories/tv2-caspar-cg-timeline-object-factory'
 import { A_B_SOURCE_INPUT_PLACEHOLDER } from '../value-objects/tv2-a-b-source-layers'
 import { Tv2FileContent } from '../value-objects/tv2-content'
-import { AssetFolderHelper } from '../helpers/AssetFolderHelper'
+import { AssetFolderHelper } from '../helpers/asset-folder-helper'
 
 const NUMBER_OF_DVE_BOXES: number = 4
 const ATEM_SUPER_SOURCE_INDEX: number = 6000
